+++ src/design/plone/contenttypes/restapi/serializers/cartella_modulistica.py
@@ -1,23 +1,26 @@
 # -*- coding: utf-8 -*-
-from Acquisition import aq_inner
-from design.plone.contenttypes.interfaces.documento import IDocumento
-from design.plone.contenttypes.restapi.serializers.dxcontent import (
-    SerializeFolderToJson,
+from .related_news_serializer import (
+    SerializeFolderToJson as RelatedNewsSerializer,
 )
+from design.plone.contenttypes.interfaces.cartella_modulistica import (
+    ICartellaModulistica,
+)
+
 from plone.restapi.interfaces import ISerializeToJson, ISerializeToJsonSummary
-from zc.relation.interfaces import ICatalog
 from zope.component import adapter, getMultiAdapter
-from zope.component import getUtility
-from zope.globalrequest import getRequest
 from zope.interface import implementer
 from zope.interface import Interface
+from Acquisition import aq_inner
+from zope.component import getUtility
 from zope.intid.interfaces import IIntIds
 from zope.security import checkPermission
+from zc.relation.interfaces import ICatalog
+from zope.globalrequest import getRequest
 
 
 @implementer(ISerializeToJson)
-@adapter(IDocumento, Interface)
-class DocumentoSerializer(SerializeFolderToJson):
+@adapter(ICartellaModulistica, Interface)
+class CartellaModulisticaSerializer(RelatedNewsSerializer):
     def get_services(self):
         """
         """
@@ -46,9 +49,8 @@ class DocumentoSerializer(SerializeFolderToJson):
         return sorted(services, key=lambda k: k["title"])
 
     def __call__(self, version=None, include_items=True):
-        if "b_size" not in self.request.form:
-            self.request.form["b_size"] = 200
-        result = super(DocumentoSerializer, self).__call__(
+        self.index = "news_uo"
+        result = super(CartellaModulisticaSerializer, self).__call__(
             version=version, include_items=include_items
         )
         result["servizi_collegati"] = self.get_services()